przeniesienie kontrolera do odpowiedniego folderu
