--- src/main/java/com/kodilla/ecommercee/OrderController.java
@@ -1,38 +1,0 @@
-package com.kodilla.ecommercee;
-
-import domain.OrderDto;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.ArrayList;
-import java.util.List;
-
-@RestController
-@RequestMapping("/orders")
-@CrossOrigin("*")
-public class OrderController {
-
-    @GetMapping
-    public List<OrderDto> getOrders() {
-        return new ArrayList<>();
-    }
-
-    @GetMapping(value = "{orderId}")
-    public OrderDto getOrder(@PathVariable long orderId) {
-        return new OrderDto();
-    }
-
-    @DeleteMapping(value = "{orderId}")
-    public void deleteOrder(@PathVariable long orderId) {
-
-    }
-
-    @PutMapping
-    public OrderDto updateOrder(@RequestBody OrderDto orderDto) {
-        return new OrderDto();
-    }
-
-    @PostMapping
-    public void createOrder(@RequestBody OrderDto orderDto) {
-
-    }
-}
