+++ src/main/java/com/kodilla/ecommercee/controller/OrderController.java
@@ -1,4 +1,38 @@
 package com.kodilla.ecommercee.controller;
 
+import com.kodilla.ecommercee.domain.OrderDto;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.ArrayList;
+import java.util.List;
+
+@RestController
+@RequestMapping("/orders")
+@CrossOrigin("*")
 public class OrderController {
+
+    @GetMapping
+    public List<OrderDto> getOrders() {
+        return new ArrayList<>();
+    }
+
+    @GetMapping(value = "{orderId}")
+    public OrderDto getOrder(@PathVariable long orderId) {
+        return new OrderDto();
+    }
+
+    @DeleteMapping(value = "{orderId}")
+    public void deleteOrder(@PathVariable long orderId) {
+
+    }
+
+    @PutMapping
+    public OrderDto updateOrder(@RequestBody OrderDto orderDto) {
+        return new OrderDto();
+    }
+
+    @PostMapping
+    public void createOrder(@RequestBody OrderDto orderDto) {
+
+    }
 }
